[Feat] change Scripts name
스크립트 이름 변경
