+++ Assets/JHT/JHTPokeType.cs
@@ -1,0 +1,11 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public enum JHTPokeType
+{
+    불,
+    물,
+    바람,
+    번개
+}
