+++ Assets/JHT/JHTSkill.cs
@@ -1,0 +1,8 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class JHTSkill : MonoBehaviour
+{
+    public float skillDamage = 5f;
+}
