+++ Assets/JHT/JHTStat.cs
@@ -1,0 +1,12 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public enum JHTStat
+{
+    체력,
+    경험치_획득량,
+    필요경험치량,
+    데미지,
+    진화
+}
