+++ Assets/JHT/JHT_TestScene/TestScene/chinaD1.prefab
@@ -1,0 +1,172 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!1001 &3923483341471139019
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    serializedVersion: 3
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: -8653719598885068355, guid: 8e19ae89c99b9764288e119683748ae9, type: 3}
+      propertyPath: m_Name
+      value: chinaD1
+      objectReference: {fileID: 0}
+    - target: {fileID: -7552582706839291426, guid: 8e19ae89c99b9764288e119683748ae9, type: 3}
+      propertyPath: m_LocalScale.x
+      value: 0.2
+      objectReference: {fileID: 0}
+    - target: {fileID: -7552582706839291426, guid: 8e19ae89c99b9764288e119683748ae9, type: 3}
+      propertyPath: m_LocalScale.y
+      value: 0.2
+      objectReference: {fileID: 0}
+    - target: {fileID: -7552582706839291426, guid: 8e19ae89c99b9764288e119683748ae9, type: 3}
+      propertyPath: m_LocalScale.z
+      value: 0.2
+      objectReference: {fileID: 0}
+    - target: {fileID: -7552582706839291426, guid: 8e19ae89c99b9764288e119683748ae9, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: -7552582706839291426, guid: 8e19ae89c99b9764288e119683748ae9, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: -7552582706839291426, guid: 8e19ae89c99b9764288e119683748ae9, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: -7552582706839291426, guid: 8e19ae89c99b9764288e119683748ae9, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: -7552582706839291426, guid: 8e19ae89c99b9764288e119683748ae9, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: -7552582706839291426, guid: 8e19ae89c99b9764288e119683748ae9, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: -7552582706839291426, guid: 8e19ae89c99b9764288e119683748ae9, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: -7552582706839291426, guid: 8e19ae89c99b9764288e119683748ae9, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: -7552582706839291426, guid: 8e19ae89c99b9764288e119683748ae9, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: -7552582706839291426, guid: 8e19ae89c99b9764288e119683748ae9, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+    m_RemovedGameObjects: []
+    m_AddedGameObjects: []
+    m_AddedComponents:
+    - targetCorrespondingSourceObject: {fileID: -8653719598885068355, guid: 8e19ae89c99b9764288e119683748ae9, type: 3}
+      insertIndex: -1
+      addedObject: {fileID: 1935281540745956684}
+    - targetCorrespondingSourceObject: {fileID: -8653719598885068355, guid: 8e19ae89c99b9764288e119683748ae9, type: 3}
+      insertIndex: -1
+      addedObject: {fileID: -5499013702259744638}
+    - targetCorrespondingSourceObject: {fileID: -8653719598885068355, guid: 8e19ae89c99b9764288e119683748ae9, type: 3}
+      insertIndex: -1
+      addedObject: {fileID: 5265950550448148196}
+    - targetCorrespondingSourceObject: {fileID: -8653719598885068355, guid: 8e19ae89c99b9764288e119683748ae9, type: 3}
+      insertIndex: -1
+      addedObject: {fileID: -489002790426076484}
+    - targetCorrespondingSourceObject: {fileID: -8653719598885068355, guid: 8e19ae89c99b9764288e119683748ae9, type: 3}
+      insertIndex: -1
+      addedObject: {fileID: 2388626464060528893}
+  m_SourcePrefab: {fileID: 4843985084834002234, guid: 8e19ae89c99b9764288e119683748ae9, type: 3}
+--- !u!1 &3572705323394426230 stripped
+GameObject:
+  m_CorrespondingSourceObject: {fileID: -8653719598885068355, guid: 8e19ae89c99b9764288e119683748ae9, type: 3}
+  m_PrefabInstance: {fileID: 3923483341471139019}
+  m_PrefabAsset: {fileID: 0}
+--- !u!95 &1935281540745956684
+Animator:
+  serializedVersion: 5
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 3572705323394426230}
+  m_Enabled: 1
+  m_Avatar: {fileID: 0}
+  m_Controller: {fileID: 9100000, guid: 0df2df1c5df189f42a25e062f435ef0d, type: 2}
+  m_CullingMode: 0
+  m_UpdateMode: 0
+  m_ApplyRootMotion: 0
+  m_LinearVelocityBlending: 0
+  m_StabilizeFeet: 0
+  m_WarningMessage: 
+  m_HasTransformHierarchy: 1
+  m_AllowConstantClipSamplingOptimization: 1
+  m_KeepAnimatorStateOnDisable: 0
+  m_WriteDefaultValuesOnDisable: 0
+--- !u!114 &-5499013702259744638
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 3572705323394426230}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 02a0a9377736cc143ac89443615c663a, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+--- !u!114 &5265950550448148196
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 3572705323394426230}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 939acd9c09e0b0144a4b384d38736ab5, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  isMyPoke: 0
+  damage: 5
+  level: 1
+  skill: []
+--- !u!114 &-489002790426076484
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 3572705323394426230}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 1df65d1a4942ad144ab94ca5297b5ba2, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  curHp: 10
+  expPoints: 10
+  isDead: 0
+--- !u!114 &2388626464060528893
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 3572705323394426230}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 260ba3d31b50817449cf89c443f75b98, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  curExp: 10
+  levelPrefabs: []
+  OnChangePrefab:
+    m_PersistentCalls:
+      m_Calls: []
