+++ Assets/JHT/JHT_TestScene/TestScene/red2.prefab
@@ -120,6 +120,12 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   curExp: 0
+  levelPrefabs:
+  - {fileID: 8151674170594283585}
+  - {fileID: 1273373693652197277, guid: 8b95c062cafe59e4aa26f786d8ca15ae, type: 3}
+  OnChangePrefab:
+    m_PersistentCalls:
+      m_Calls: []
 --- !u!114 &5493569055312173983
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -134,6 +140,7 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   curHp: 0
   expPoints: 10
+  isDead: 0
 --- !u!114 &4761624671728559483
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -148,5 +155,5 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   isMyPoke: 0
   damage: 3
-  level: 1
+  level: 2
   skill: []
