+++ Assets/JHT/JHT_TestScene/TestScene/red3.prefab
@@ -1,0 +1,122 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!1001 &1606695998538251808
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    serializedVersion: 3
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: -8653719598885068355, guid: 89ea2beee5c1706459f683c8e2fda03a, type: 3}
+      propertyPath: m_Name
+      value: red3
+      objectReference: {fileID: 0}
+    - target: {fileID: -7552582706839291426, guid: 89ea2beee5c1706459f683c8e2fda03a, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: -7552582706839291426, guid: 89ea2beee5c1706459f683c8e2fda03a, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: -7552582706839291426, guid: 89ea2beee5c1706459f683c8e2fda03a, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: -7552582706839291426, guid: 89ea2beee5c1706459f683c8e2fda03a, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: -7552582706839291426, guid: 89ea2beee5c1706459f683c8e2fda03a, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: -7552582706839291426, guid: 89ea2beee5c1706459f683c8e2fda03a, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: -7552582706839291426, guid: 89ea2beee5c1706459f683c8e2fda03a, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: -7552582706839291426, guid: 89ea2beee5c1706459f683c8e2fda03a, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: -7552582706839291426, guid: 89ea2beee5c1706459f683c8e2fda03a, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: -7552582706839291426, guid: 89ea2beee5c1706459f683c8e2fda03a, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+    m_RemovedGameObjects: []
+    m_AddedGameObjects: []
+    m_AddedComponents:
+    - targetCorrespondingSourceObject: {fileID: -8653719598885068355, guid: 89ea2beee5c1706459f683c8e2fda03a, type: 3}
+      insertIndex: -1
+      addedObject: {fileID: -233127128852138717}
+    - targetCorrespondingSourceObject: {fileID: -8653719598885068355, guid: 89ea2beee5c1706459f683c8e2fda03a, type: 3}
+      insertIndex: -1
+      addedObject: {fileID: 4814124939929986014}
+    - targetCorrespondingSourceObject: {fileID: -8653719598885068355, guid: 89ea2beee5c1706459f683c8e2fda03a, type: 3}
+      insertIndex: -1
+      addedObject: {fileID: -6344021220332474178}
+  m_SourcePrefab: {fileID: 4843985084834002234, guid: 89ea2beee5c1706459f683c8e2fda03a, type: 3}
+--- !u!1 &1273373693652197277 stripped
+GameObject:
+  m_CorrespondingSourceObject: {fileID: -8653719598885068355, guid: 89ea2beee5c1706459f683c8e2fda03a, type: 3}
+  m_PrefabInstance: {fileID: 1606695998538251808}
+  m_PrefabAsset: {fileID: 0}
+--- !u!114 &-233127128852138717
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1273373693652197277}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 260ba3d31b50817449cf89c443f75b98, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  curExp: 0
+  levelPrefabs:
+  - {fileID: 1273373693652197277}
+  OnChangePrefab:
+    m_PersistentCalls:
+      m_Calls: []
+--- !u!114 &4814124939929986014
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1273373693652197277}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 1df65d1a4942ad144ab94ca5297b5ba2, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  curHp: 10
+  expPoints: 10
+  isDead: 0
+--- !u!114 &-6344021220332474178
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1273373693652197277}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 939acd9c09e0b0144a4b384d38736ab5, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  isMyPoke: 0
+  damage: 10
+  level: 3
+  skill: []
