--- Assets/JHT/PokeType.cs
@@ -1,11 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public enum PokeType
-{
-    불,
-    물,
-    바람,
-    번개
-}
