--- Assets/JHT/Skill.cs
@@ -1,8 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class Skill : MonoBehaviour
-{
-    public float skillDamage = 5f;
-}
