--- Assets/JHT/Stat.cs
@@ -1,12 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public enum Stat
-{
-    체력,
-    경험치_획득량,
-    필요경험치량,
-    데미지,
-    진화
-}
